+++ wp_api/src/wp_com/client.rs
@@ -32,6 +32,7 @@ use super::endpoint::{
         StatsFileDownloadsRequestBuilder, StatsFileDownloadsRequestExecutor,
     },
     stats_insights_endpoint::{StatsInsightsRequestBuilder, StatsInsightsRequestExecutor},
+    stats_post_views_endpoint::{StatsPostViewsRequestBuilder, StatsPostViewsRequestExecutor},
     stats_referrers_endpoint::{StatsReferrersRequestBuilder, StatsReferrersRequestExecutor},
     stats_region_views_endpoint::{
         StatsRegionViewsRequestBuilder, StatsRegionViewsRequestExecutor,
@@ -94,6 +95,7 @@ pub struct WpComApiRequestBuilder {
     stats_emails_summary: Arc<StatsEmailsSummaryRequestBuilder>,
     stats_devices_platform: Arc<StatsDevicesPlatformRequestBuilder>,
     stats_devices_screensize: Arc<StatsDevicesScreensizeRequestBuilder>,
+    stats_post_views: Arc<StatsPostViewsRequestBuilder>,
     stats_referrers: Arc<StatsReferrersRequestBuilder>,
     stats_subscribers: Arc<StatsSubscribersRequestBuilder>,
     stats_region_views: Arc<StatsRegionViewsRequestBuilder>,
@@ -141,6 +143,7 @@ impl WpComApiRequestBuilder {
             stats_emails_summary,
             stats_devices_platform,
             stats_devices_screensize,
+            stats_post_views,
             stats_referrers,
             stats_subscribers,
             stats_region_views,
@@ -199,6 +202,7 @@ pub struct WpComApiClient {
     stats_emails_summary: Arc<StatsEmailsSummaryRequestExecutor>,
     stats_devices_platform: Arc<StatsDevicesPlatformRequestExecutor>,
     stats_devices_screensize: Arc<StatsDevicesScreensizeRequestExecutor>,
+    stats_post_views: Arc<StatsPostViewsRequestExecutor>,
     stats_referrers: Arc<StatsReferrersRequestExecutor>,
     stats_subscribers: Arc<StatsSubscribersRequestExecutor>,
     stats_region_views: Arc<StatsRegionViewsRequestExecutor>,
@@ -247,6 +251,7 @@ impl WpComApiClient {
             stats_emails_summary,
             stats_devices_platform,
             stats_devices_screensize,
+            stats_post_views,
             stats_referrers,
             stats_subscribers,
             stats_region_views,
@@ -288,6 +293,7 @@ api_client_generate_endpoint_impl!(WpComApi, stats_devices_browser);
 api_client_generate_endpoint_impl!(WpComApi, stats_emails_summary);
 api_client_generate_endpoint_impl!(WpComApi, stats_devices_platform);
 api_client_generate_endpoint_impl!(WpComApi, stats_devices_screensize);
+api_client_generate_endpoint_impl!(WpComApi, stats_post_views);
 api_client_generate_endpoint_impl!(WpComApi, stats_referrers);
 api_client_generate_endpoint_impl!(WpComApi, stats_subscribers);
 api_client_generate_endpoint_impl!(WpComApi, stats_region_views);